@@ -2,9 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_reactive_ble/flutter_reactive_ble.dart';
 
 class BleStatusScreen extends StatelessWidget {
-  const BleStatusScreen({@required this.status, Key key})
-      : assert(status != null),
-        super(key: key);
+  const BleStatusScreen({required this.status, Key? key}) : super(key: key);
 
   final BleStatus status;
 
